backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.12
python-docx>=1.1.0
pymupdf>=1.24.0
boto3>=1.35.0
pytesseract>=0.3.10
Pillow>=10.0.0
